+++ src/markehme/factionsplus/Cmds/CmdJail.java
@@ -3,6 +3,7 @@ package markehme.factionsplus.Cmds;
 import org.bukkit.*;
 
 import markehme.factionsplus.*;
+import markehme.factionsplus.config.*;
 import markehme.factionsplus.config.sections.Section_Jails;
 
 import com.massivecraft.factions.cmd.FCommand;
@@ -35,7 +36,11 @@ public class CmdJail extends FCommand {
 		}
 		if (Section_Jails.canJailUnjail( fme ))
 		{
-			FactionsPlusJail.sendToJail( playerToJail, Utilities.getOnlinePlayerExact(fme), -1 );
+			if ( ( !Config._jails.canJailOnlyIfIssuerIsInOwnTerritory._ ) || ( fme.isInOwnTerritory() ) ) {
+				FactionsPlusJail.sendToJail( playerToJail, Utilities.getOnlinePlayerExact( fme ), -1 );
+			} else {
+				fme.sendMessage( ChatColor.RED + "You must be in your own faction territory to jail someone." );
+			}
 			return;
 		}
 		fme.sendMessage(ChatColor.RED+ "No permission to jail!" );